Revert "add PLATFORM_X86 define for code intended for both i386 and x86_64"
This reverts commit b2ab839a00fb79dc30ed74f35cdd3e7dc8e416e4.

+++ public/steam/steamtypes.h
@@ -77,11 +77,11 @@ typedef long int lint64;
 typedef unsigned long int ulint64;
 
 #ifdef X64BITS
-typedef int64 intp;
-typedef uint64 uintp;
+typedef int64_t intp;
+typedef uint64_t uintp;
 #else
-typedef int64 intp;
-typedef uint64 uintp;
+typedef int32_t intp;
+typedef uint32_t uintp;
 #endif
 
 #endif // else _WIN32
